requirements.txt
discord.py>=1.5.1
numpy>=1.18.2